requirements-redaction.txt
watchdog>=3.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
python-docx>=1.1.0
spacy>=3.7.0
# After installing, also run:
#   python -m spacy download en_core_web_lg
